Move Spring/JDBC related classes to jdbc module. Fixes #787
