+++ configurations/hibernate-jpa/src/main/java/io/micronaut/configuration/hibernate/jpa/HibernateTransactionManagerFactory.java
@@ -17,11 +17,10 @@
 package io.micronaut.configuration.hibernate.jpa;
 
 import io.micronaut.context.annotation.*;
-import io.micronaut.spring.tx.datasource.DataSourceTransactionManagerFactory;
+import io.micronaut.jdbc.spring.DataSourceTransactionManagerFactory;
 import org.hibernate.SessionFactory;
 import org.springframework.orm.hibernate5.HibernateTransactionManager;
 
-import javax.inject.Singleton;
 import javax.sql.DataSource;
 
 /**
