+++ jdbc/build.gradle
@@ -1,3 +1,8 @@
 dependencies {
+    compileOnly project(":inject-java")
+    compileOnly project(":spring")
+    compileOnly dependencyModuleVersion("spring", "spring-jdbc")
+    compileOnly dependencyModuleVersion("spring", "spring-tx")
+
     compile project(":inject")
 }
