+++ spring/build.gradle
@@ -5,7 +5,6 @@ dependencies {
     compile project(":inject")
     compileOnly project(":inject-java")
     compileOnly dependencyModuleVersion("spring", "spring-jdbc")
-    compileOnly project(":jdbc")
     compile project(":aop")
 
     testImplementation project(':runtime')
